begin implementing dark mode
