+++ src/components/ApplicationSection.tsx
@@ -7,7 +7,7 @@ import { useAppContext } from "../context/AppContext";
 import { RiFolderAddLine } from "react-icons/ri";
 import { Filter } from "./Filter";
 import { LineDesign } from "./LineDesign";
-import { IApplicationModel } from "../models/models";
+import { IApplicationModel } from "../interfaces/interfaces";
 
 export const ApplicationSection = () => {
   const [filter, setFilter] = useState<string>("all");
@@ -84,7 +84,7 @@ const Wrapper = styled.div`
     box-shadow: 4px 4px 8px rgba(0, 0, 0, 0.1);
     .my-applications {
       margin-top: 0.5rem;
-      font-family: "Josefin Slab", serif;
+      font-family: "Poppins", sans-serif;
     }
     .button-wrapper {
       padding: 1rem;
