+++ src/components/Filter.tsx
@@ -59,7 +59,7 @@ const Wrapper = styled.form`
   }
 
   .filter-btn {
-    font-family: "Playfair Display", serif;
+    font-family: "Poppins", sans-serif;
     position: relative;
     z-index: 1;
     height: 2rem;
