+++ src/pages/EditAppPage.tsx
@@ -306,7 +306,7 @@ const Wrapper = styled.main`
       border: 2px solid rgba(0, 0, 0, 0.3);
       border-radius: 2px;
       background-color: rgba(215, 210, 255, 0.5);
-      font-family: "Playfair Display", serif;
+      font-family: "Poppins", sans-serif;
       font-weight: 700;
       &:hover {
         cursor: pointer;
