+++ src/pages/LoginPage.tsx
@@ -205,7 +205,7 @@ const Wrapper = styled.main`
       top: 0.5rem;
       text-align: center;
       color: red;
-      font-family: "Josefin Slab", serif;
+      font-family: "Poppins", sans-serif;
     }
     .form-input {
       margin-bottom: 1rem;
@@ -228,7 +228,7 @@ const Wrapper = styled.main`
       border: 2px solid rgba(0, 0, 0, 0.3);
       border-radius: 2px;
       background-color: rgba(215, 210, 255, 0.5);
-      font-family: "Playfair Display", serif;
+      font-family: "Poppins", sans-serif;
       font-weight: 700;
       font-size: 1.1rem;
       &:hover {
